Add option to adjust model climo to look like real climo
The `—adjust-model-climo` option will adjust the model climatology (`ym`) to look like the real 30-year climo. It does this by taking the different between the real 30-year climo and `xm` (the mean obs within the reforecast period), then adding that difference to `ym`. From then on in the regression, `ym` is more representative of the 30-year climo period.