Limit rxy to be less than or equal to 0.9
When rxy is really high (eg. higher terrain of the west it can be near 1 sometimes), it causes a noisy k and hence a noisy ebest, which makes for a splotchy map (see noaa-nws-cpc/mpp-driver#35).

Limiting rxy to not be above 0.9 makes those splotchy probabilities somewhat lower (instead of near 100%).